avoid pessimistic busy render loop
I noticed that running `sp` would spin 100% cpu for both `sp`
and my terminal emulator.  I ran this down to this bit of code,
which would force a full screen render every 100ms.

What seems to work for me is making this only return a Refresh
action if there is a pending refresh.