ignore wake events with no queued events
Collection of events from the mpsc queue can overtake the corresponding
wake event.  If this happens, returning early can cause busy looping
as we continuously send and receive render events.

Instead, if we recieve a wake event and there is nothing queued, ignore
the wake event and go back to waiting.  Normally this will only happen
if the wake event was immediately delivered, which means our wait delay
is probably still the right length of time to wait.