use attributes as base for search match highlights
When computing the attributes for search matches, start from the content
attributes, rather than the default attributes.  This means that attributes
other than the ones that are used for highlighting still apply to matches.
This is most important for line drawing mode, but it's useful for other things
(like underline or italics) too.