wait for a complete line in Direct mode
The direct mode does not re-render output lines. It treats lines as immutable.
However, the last line in output is mutable if the stream does not write "\n".
Detect that and avoid writing the mutable line.

`stream_example.rs` is changed to demonstrate the issue. Before this change,
`MODE=direct cargo run --example streams_example` will not print the numbers
of lines correctly.