implement new file reader
Implement a new reader for on-disk files, that reads the data out of the file
rather than mmapping it.

Reads to the file for display and search go through an LRU cache of 1MB blocks.

If a file change is detected, then depending on whether or not the file
has been appended to or rewritten, streampager will load the remaining data
or flush its caches and reload the file.

The heuristic for append-vs-reload is whether the last 4k of the file has
changed or not, or if any time we try to parse a line we find a newline
in the middle of the line.

This replaces the mmap implementation, giving a better experience for looking
at files on disk.

The mmap implementation is retained in case it will be useful, but for now is
unused.

Fixes #8
Fixes #9