file: make File::drop stop background threads
The direct motivation is to solve the "exit early" issue with "[loading
paused]" state. That can be done by

    impl Drop for File {
        fn drop(&mut self) {
            // Unblock paused loading thread.
            file.set_needed_lines(usize::max_value());
        }
    }

However that is not obvious what's going on.

This patch adds a new field to FileMeta to indicate the File was dropped.
The field is then checked by the background threads when they are "idle".

With this change, `cargo run --example streams_example -- inf` no longer
freezes after the user presses 'q'.