add default wrapping mode to the config
Add `wrapping_mode` to the config, allowing callers to decide on the
default wrapping mode.  Users can still switch wrapping mode using
the `\` key, this just sets the default.

When `-X` is specified, default to line wrapping to match direct mode
behaviour.

Note that no environment variable is added to control this.  We will add a
config file to prevent proliferation of env vars.