ensure file loading is stopped when all users of the file have finished
The `File` object may be legitimately cloned; this is commonly done with error
files which are referenced from multiple screens.

The current `Drop` implementation for `File` stops the background thread when
the *first* copy of `File` is dropped.  This means files can suddenly stop loading
even when they are in use.

Only stop the file loading thread if *all* of the `File` clones have been
dropped, by using a drop guard that is shared amongst all `File` instances.