update buffer before releasing newlines lock
The parsing of newlines can race with things that are rendering lines (in
particular, the `direct`-mode renderer).  We need to ensure that the
buffer has been updated *before* we release the lock with the new newlines
parsed, so that anything that is looking at the lines can successfully
read the newly added lines.