Refactor to use our own Error type in the library api
Using anyhow in a library's public interface is not recommended mainly
as it makes it tough to integrate with other libraries or existing
applications which may not use anyhow themselves, chiefly due to
anyhow::Error not implementing std::error::Error itself.

This commit instead refactors the code to use several thiserror[1]-derived
Error types and completely eliminates anyhow from library code.

It is still used as the output error library in the sp and spp
executables, though.

[1]: https://docs.rs/thiserror