keep event types private
Providing wrapped versions of the channel send errors in the
main Error type causes the event types to be leaked from the crate.
This is because the error contains the unsent message, so that it can
be recovered.

We would prefer these types don't get leaked, in case we want to change
them in the future.  They're internal, so it's unlikely a client
receiving the error would be able to do anything useful with them
either.

Change the Error type to discard this extra information, and make these
types private to the crate again.