stop loading files when the original File is dropped
File references can be cloned and kept alive by threads (e.g. the search
thread).  This means when a file is paused, we won't close the pipe when
the pager is exiting if search is open.  This matters when streampager
is embedded within a process, as that process may be waiting for the
pipe to close before it exits.

Change so that the original `File` object is special: when it is dropped
the file is considered closed (even if there are other references to
it).