render wrapped lines as a single line
When the rendering mode is set to wrapping on grapheme boundaries, and line
numbers are not enabled, render long lines together as a single line, allowing
the terminal to wrap.  This will let the terminal understand that this is a
long line, and allow copying of text to work for the wrapped line.