subtract the bottom line when scrolling up direct output
When exiting direct mode and switching to paging, we scroll the output up so
that we don't clobber earlier lines of output.  However, during direct mode the
cursor occupied the bottom line of the screen, so we must push up the content
by one less line than the height of the screen, otherwise we'll push it too far
and end up duplicating the first line of the file in the terminal scrollback.