add startup_poll_input config option
Add a `startup_poll_input` config option (default: true).   If set to false,
streampager will not poll input during the delayed or direct mode startup
phase.  This means input typed before the pager exits early for these modes
will remain in the terminal input buffer (e.g. for a subsequent command
passed to the shell).

Note that a side-effect of this is is that we also miss SIGWINCH events, as
they come from the termiz terminal on the same event stream.

Fixes #39.